refactor: 도메인과 뷰 로직을 각각 domain/, view/ 하위 경로에서 관리하도록 수정
